feat: reject additional arguments that are managed via explicit runner builder methods
